Mueve viewsets y serializadores al módulo API
