+++ config/settings/production.py
@@ -75,7 +75,7 @@ LOGGING = {
             'backupCount': 5,
             'maxBytes': 1024*1024*100,  # 100MB
             'formatter': 'verbose',
-            'filename': str(ROOT_DIR / "errors.log"),
+            'filename': str(ROOT_DIR / "errors.log"), # noqa F405
         },
     },
     'loggers': {
