+++ sistemita/core/templates/base.html
@@ -58,38 +58,38 @@
     </div>
 
     <script>
-      {% block extra_js %}{% endblock %}
+    {% block extra_js %}{% endblock %}
 
-      $('#id_provincia').change(function(){
-      var provincia = $('#id_provincia').val();
-      var select_distrito = $('#id_distrito');
-      var select_localidad = $('#id_localidad');
+    $('#id_provincia').change(function(){
+	var provincia = $('#id_provincia').val();
+	var select_distrito = $('#id_distrito');
+	var select_localidad = $('#id_localidad');
 
-      // Limpio ditrito y localidad
-      select_distrito.children().remove().end();
-      select_distrito.append($("<option />").text('----'));
-      select_localidad.children().remove().end();
-      select_localidad.append($("<option />").text('----'));
+	// Limpio ditrito y localidad
+	select_distrito.children().remove().end();
+	select_distrito.append($("<option />").text('----'));
+	select_localidad.children().remove().end();
+	select_localidad.append($("<option />").text('----'));
 
-      // Actualizo
-      $.getJSON("{% url 'core:distrito-list' %}?provincia=" + provincia, function(distritos) {
-      $.each(distritos, function(index, distrito) {
-      select_distrito.append($("<option />").val(distrito.id).text(distrito.nombre));
-      });
-      });
-      });
+	// Actualizo
+	$.getJSON("{% url 'api:distrito-list' %}?provincia=" + provincia, function(distritos) {
+	    $.each(distritos, function(index, distrito) {
+		select_distrito.append($("<option />").val(distrito.id).text(distrito.nombre));
+	    });
+	});
+    });
 
-      $('#id_distrito').change(function(){
-      var distrito = $('#id_distrito').val();
-      var select_localidad = $('#id_localidad');
-      select_localidad.children().remove().end();
-      select_localidad.append($("<option />").text('----'));
-      $.getJSON("{% url 'core:localidad-list' %}?distrito=" + distrito, function(localidades) {
-      $.each(localidades, function(index, localidad) {
-      select_localidad.append($("<option />").val(localidad.id).text(localidad.nombre));
-      });
-      });
-      });
+    $('#id_distrito').change(function(){
+	var distrito = $('#id_distrito').val();
+	var select_localidad = $('#id_localidad');
+	select_localidad.children().remove().end();
+	select_localidad.append($("<option />").text('----'));
+	$.getJSON("{% url 'api:localidad-list' %}?distrito=" + distrito, function(localidades) {
+	    $.each(localidades, function(index, localidad) {
+		select_localidad.append($("<option />").val(localidad.id).text(localidad.nombre));
+	    });
+	});
+    });
     </script>
 
   </body>
