+++ sistemita/core/views/clientes.py
@@ -13,13 +13,9 @@ from django.urls import reverse, reverse_lazy
 from django.views.generic import DeleteView, DetailView, ListView
 from django.views.generic.edit import CreateView, UpdateView
 
-# Django REST Framework
-from rest_framework import mixins, permissions, viewsets
-
 # Sistemita
 from sistemita.core.forms.clientes import ClienteForm
 from sistemita.core.models.cliente import Cliente
-from sistemita.core.serializers import ClienteSerializer
 from sistemita.core.utils.strings import (
     MESSAGE_403,
     MESSAGE_SUCCESS_CREATED,
@@ -27,14 +23,6 @@ from sistemita.core.utils.strings import (
     MESSAGE_SUCCESS_UPDATE,
 )
 from sistemita.core.views.home import error_403
-
-
-class ClienteViewSet(mixins.ListModelMixin, mixins.RetrieveModelMixin, viewsets.GenericViewSet):
-    """Cliente View set."""
-
-    queryset = Cliente.objects.all()
-    permission_classes = (permissions.IsAuthenticated,)
-    serializer_class = ClienteSerializer
 
 
 class ClienteListView(PermissionRequiredMixin, SuccessMessageMixin, ListView):
