+++ sistemita/core/views/facturascliente.py
@@ -19,10 +19,10 @@ from django_filters.views import FilterView
 
 # Sistemita
 from sistemita.accounting.models.cobranza import Cobranza, CobranzaFactura
+from sistemita.core.constants import TIPOS_FACTURA_IMPORT
 from sistemita.core.filters import FacturaFilterSet
 from sistemita.core.forms.clientes import FacturaForm
 from sistemita.core.models.cliente import Factura
-from sistemita.core.constants import TIPOS_FACTURA_IMPORT
 from sistemita.core.utils.export import export_excel
 from sistemita.core.utils.strings import (
     _MESSAGE_SUCCESS_CREATED,
