+++ sistemita/core/views/mediopago.py
@@ -12,13 +12,9 @@ from django.urls import reverse, reverse_lazy
 from django.views.generic import DeleteView, DetailView, ListView
 from django.views.generic.edit import CreateView, UpdateView
 
-# Django REST Framework
-from rest_framework import mixins, permissions, viewsets
-
 # Sistemita
 from sistemita.core.forms.mediospago import MedioPagoForm
 from sistemita.core.models.mediopago import MedioPago
-from sistemita.core.serializers import MedioPagoSerializer
 from sistemita.core.utils.strings import (
     MESSAGE_403,
     MESSAGE_SUCCESS_CREATED,
@@ -26,14 +22,6 @@ from sistemita.core.utils.strings import (
     MESSAGE_SUCCESS_UPDATE,
 )
 from sistemita.core.views.home import error_403
-
-
-class MedioPagoViewSet(mixins.ListModelMixin, mixins.RetrieveModelMixin, viewsets.GenericViewSet):
-    """Medio de pago view set."""
-
-    queryset = MedioPago.objects.all()
-    serializer_class = MedioPagoSerializer
-    permission_classes = (permissions.IsAuthenticated,)
 
 
 class MedioPagoListView(PermissionRequiredMixin, SuccessMessageMixin, ListView):
